Rearrange files and functions to make for a more readable app
