--- giudice/__init__.py
@@ -1,32 +1,0 @@
-# import os
-
-# from flask import Flask
-
-
-# def create_app(test_config=None):
-#     # create and configure the app
-#     app = Flask(__name__, instance_relative_config=True)
-#     app.config.from_mapping(
-#         SECRET_KEY='dev',
-#         DATABASE=os.path.join(app.instance_path, 'flaskr.sqlite'),
-#     )
-
-#     if test_config is None:
-#         # load the instance config, if it exists, when not testing
-#         app.config.from_pyfile('config.py', silent=True)
-#     else:
-#         # load the test config if passed in
-#         app.config.from_mapping(test_config)
-
-#     # ensure the instance folder exists
-#     try:
-#         os.makedirs(app.instance_path)
-#     except OSError:
-#         pass
-
-#     # a simple page that says hello - use it to make sure you didn't crash the whole app :,)
-#     @app.route('/hello')
-#     def hello():
-#         return 'Hello, World!'
-
-#     return app
