+++ scripts/run-tests.sh
@@ -1,3 +1,4 @@
 #!/bin/bash
 
 python -c 'from tests.main_tests import *; print run_tests()'
+python -c 'from tests.projects_tests import *; print run_tests()'
